requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
diskcache==5.6.3
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.2
frozenlist==1.4.1
greenlet==3.1.1
groq==0.11.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.4
langchain-core==0.3.12
langchain-text-splitters==0.3.0
langsmith==0.1.136
llama_cpp_python==0.3.1
MarkupSafe==3.0.2
multidict==6.1.0
numpy==1.26.4
orjson==3.10.9
packaging==24.1
pillow==11.0.0
playwright==1.48.0
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
pyee==12.0.0
PyMuPDF==1.24.12
PyPDF2==3.0.1
pytesseract==0.3.13
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.40.0
tenacity==9.0.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
yarl==1.15.5
